requirements.txt
streamlit==1.19.0
numpy==1.24.0
matplotlib==3.7.1
tensorflow==2.14.0
Pillow==10.0.0